% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitModel.R
\name{fit}
\alias{fit}
\title{Fit end-to-end ICAR model}
\usage{
fit(
  dtBasicSetUp_one_simu,
  dtYfullTable_one_simu,
  dtNtarget_one_simu,
  dtyObs_by_loc_one_simu,
  tau_prior,
  sigma_theta_prior,
  theta_prior,
  alpha,
  tau2,
  prec_theta,
  prev_ub = 0.2,
  sigma_mu = 10,
  sigma_loading = 10,
  SD_HalfCauchy_Prior = 10,
  seedLoadingMat = NaN,
  seedTheta = NaN,
  seedCovTheta = NaN,
  monitors = c("N", "pesudo_logit_prev", "sigma2.phi", "resPhi", "beta0",
    "loading_mat_upper_tri"),
  nchain = 3,
  iter = 68000,
  nburnin = 50000,
  nthin = 100,
  setSeed = FALSE
)
}
\arguments{
\item{dtBasicSetUp_one_simu}{Example of data - writing some descriptions here}
}
\value{
Posterior samples of ICAR model paramters
}
\description{
Fit end-to-end ICAR model
}
\examples{
fit(dtBasicSetUp_one_simu,
    dtYfullTable_one_simu,
    dtNtarget_one_simu,
    dtyObs_by_loc_one_simu,
    tau_prior = c(inv_tau2  ~ T(dt(mu = 0, tau = sd_spRE_prior, df = 1), 0, ),"tau2 <- 1/inv_tau2"),
    igma_theta_prior = c(sigma_theta ~ T(dt(mu = 0, tau = sd_theta_prior, df = 1), 0, )),
    theta_prior = c(theta[i] ~ dnorm(mu_theta, sd = sigma_theta),"loading_latent_factor[i,1:J] <- theta[i] * t(loading_mat_upper_tri[1:J,1])"),
    alpha = c(0.05, 0.03, 0.06),
    tau2 = c(0.5, 0.1, 0.2),
    prec_theta = c(10, 20, 50),
    prev_ub = 0.2,
    sigma_mu = 10,
    sigma_loading = 10,
    SD_HalfCauchy_Prior = 10,
    seedLoadingMat = c(123, 234, 345),
    seedTheta = c(123, 234, 345),
    seedCovTheta=NaN,
    monitors = c("N","pesudo_logit_prev", "sigma2.phi", "resPhi", "beta0","loading_mat_upper_tri"),
    nchain = 3,
    iter = 68000,
    nburnin = 50000,
    nthin = 100,
    setSeed=9)
}
