% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelGraph.R
\name{hyper2Likelihood}
\alias{hyper2Likelihood}
\title{Full likelihood of data and prior and hyper}
\usage{
hyper2Likelihood(tau_prior, sigma_theta_prior, theta_prior)
}
\arguments{
\item{tau_prior}{hyper of phi}

\item{sigma_theta_prior}{hyper of theta}

\item{theta_prior}{prior theta}
}
\value{
return the full likelihood of ICAR model
}
\description{
Full likelihood of data and prior and hyper
}
\examples{
tau_prior <- c(inv_tau2  ~ T(dt(mu = 0, tau = sd_spRE_prior, df = 1), 0, ),"tau2 <- 1/inv_tau2")
sigma_theta_prior <- c(sigma_theta ~ T(dt(mu = 0, tau = sd_theta_prior, df = 1), 0, ))
theta_prior <- c(theta[i] ~ dnorm(mu_theta, sd = sigma_theta), "loading_latent_factor[i,1:J] <- theta[i] * t(loading_mat_upper_tri[1:J,1])")
model_ll <- hyper2Likelihood(tau_prior, sigma_theta_prior, theta_prior)
}
